Tried to automate the downloading from Cronometer
Tried with Pyppeteer: Package is porly supported
Playwright: the document layout is more complicated than I really want
to deal with.
Puppeteer: closest, but just times out, and there's no sensible way to
get the downloaded files into a sensible location.

Meanwhile, manually downloading the files takes seconds, so it's
probably not worth the time it would take to automate this task.

started developing the playwright utility to export Cronometer data.

intial work on playwright.
Is this worth it?